requirements.txt
keyboard==0.13.5
colorama==0.4.6
python-dotenv==1.0.1
pyautogui==0.9.54